// src/store/useCart/enum.ts
export enum ToastMessages {
  ADDED_SUCCESSFUL = "Product added successful",
  ADDED_MAX_QTY = "Max quantity added",
  REMOVED_SUCCESSFUL = "Product removed successful",
  REMOVED_ERROR = "It is not possible to remove more items.",
  DELETED_SUCCESSFUL = "Product deleted successful",
}

export enum ToastTypes {
  SUCCESS = "success",
  INFO = "info",
  ERROR = "error",
}